Added format_time_since()
Retuns a readable & formatted time from a datetime to now
Note: the further away the datetime is, the more inconsistant the time is. Why? Because leap years, YEAH, FUCKING LEAP YEARS ARE CRUSHING MY DREAMS OF HAVING A PERFECT TIME SYSTEM BECAUSE I'M USING AVERAGE DAYS IN A MONTH AND A YEAR SO IT GETS MORE AND MORE UNCERTAIN THE OLDER THE DATETIME IS.
use at your own risk
but tbh if it's for followed since it works fine